requirements.txt
mysql-connector-python==8.0.13
pyinstaller==3.4
tabulate==0.8.2
